code review fixes: revert the int->uint change, bring back the #if DEBUG

+++ src/System.Private.CoreLib/shared/System/Buffers/TlsOverPerCoreLockedStacksArrayPool.cs
@@ -135,13 +135,13 @@ namespace System.Buffers
                 }
 
                 // No buffer available.  Allocate a new buffer with a size corresponding to the appropriate bucket.
-                buffer = GC.AllocateUninitializedArray<T>((uint)_bucketArraySizes[bucketIndex]);
+                buffer = GC.AllocateUninitializedArray<T>(_bucketArraySizes[bucketIndex]);
             }
             else
             {
                 // The request was for a size too large for the pool.  Allocate an array of exactly the requested length.
                 // When it's returned to the pool, we'll simply throw it away.
-                buffer = GC.AllocateUninitializedArray<T>((uint)minimumLength);
+                buffer = GC.AllocateUninitializedArray<T>(minimumLength);
             }
 
             if (log.IsEnabled())
